+++ json_annotation/lib/src/json_serializable.g.dart
@@ -29,6 +29,8 @@ JsonSerializable _$JsonSerializableFromJson(
         'generic_argument_factories',
         'ignore_unannotated',
         'include_if_null',
+        'union_discriminator',
+        'union_rename',
       ],
     );
     final val = JsonSerializable(
@@ -61,6 +63,14 @@ JsonSerializable _$JsonSerializableFromJson(
         'create_per_field_to_json',
         (v) => v as bool?,
       ),
+      unionDiscriminator: $checkedConvert(
+        'union_discriminator',
+        (v) => v as String?,
+      ),
+      unionRename: $checkedConvert(
+        'union_rename',
+        (v) => $enumDecodeNullable(_$FieldRenameEnumMap, v),
+      ),
     );
     return val;
   },
@@ -77,6 +87,8 @@ JsonSerializable _$JsonSerializableFromJson(
     'includeIfNull': 'include_if_null',
     'genericArgumentFactories': 'generic_argument_factories',
     'createPerFieldToJson': 'create_per_field_to_json',
+    'unionDiscriminator': 'union_discriminator',
+    'unionRename': 'union_rename',
   },
 );
 
@@ -96,6 +108,8 @@ Map<String, dynamic> _$JsonSerializableToJson(JsonSerializable instance) =>
       'generic_argument_factories': instance.genericArgumentFactories,
       'ignore_unannotated': instance.ignoreUnannotated,
       'include_if_null': instance.includeIfNull,
+      'union_discriminator': instance.unionDiscriminator,
+      'union_rename': _$FieldRenameEnumMap[instance.unionRename],
     };
 
 const _$FieldRenameEnumMap = {