+++ json_serializable/lib/src/encoder_helper.dart
@@ -11,6 +11,7 @@ import 'helper_core.dart';
 import 'type_helpers/generic_factory_helper.dart';
 import 'type_helpers/json_converter_helper.dart';
 import 'unsupported_type_error.dart';
+import 'utils.dart';
 
 mixin EncodeHelper implements HelperCore {
   String _fieldAccess(FieldElement field) => '$_toJsonParamName.${field.name!}';
@@ -72,7 +73,6 @@ mixin EncodeHelper implements HelperCore {
     final buffer = StringBuffer(
       'abstract final class _\$${element.name!.nonPrivate}JsonKeys {',
     );
-    // ..write('static const _\$${element.name.nonPrivate}JsonKeys();');
 
     for (final field in accessibleFieldSet) {
       buffer.writeln(
@@ -89,6 +89,23 @@ mixin EncodeHelper implements HelperCore {
   Iterable<String> createToJson(Set<FieldElement> accessibleFields) sync* {
     assert(config.createToJson);
 
+    final expressionBody = element.isSealed
+        ? _createSealedFunctionExpressionBody()
+        : _createFieldMapFunctionExpressionBody(accessibleFields);
+
+    yield _createToJsonFunctionSignature(expressionBody);
+  }
+
+  /// Creates the function signature around [functionExpressionBody]
+  /// that will be used to serialize the class.
+  ///
+  /// For example:
+  ///
+  /// ```dart
+  /// Map<String, dynamic> _$ExampleClassToJson(ExampleClass instance) =>
+  ///       /* expression body here */;
+  /// ```
+  String _createToJsonFunctionSignature(String functionExpressionBody) {
     final buffer = StringBuffer();
 
     final functionName =
@@ -100,25 +117,82 @@ mixin EncodeHelper implements HelperCore {
 
     if (config.genericArgumentFactories) _writeGenericArgumentFactories(buffer);
 
-    buffer
-      ..write(') ')
-      ..writeln('=> <String, dynamic>{')
-      ..writeAll(
-        accessibleFields.map((field) {
-          final access = _fieldAccess(field);
+    buffer.write(') => $functionExpressionBody;');
 
-          final keyExpression = safeNameAccess(field);
-          final valueExpression = _serializeField(field, access);
+    return buffer.toString();
+  }
 
-          final maybeQuestion = _canWriteJsonWithoutNullCheck(field) ? '' : '?';
+  /// Creates expression body for a function that serializes a union class.
+  ///
+  /// For example:
+  /// ```dart
+  /// switch (instance) {
+  ///   final FirstSubtype instance => {
+  ///       'type': 'FirstSubtype',
+  ///       ..._$FirstSubtypeToJson(instance),
+  ///     },
+  ///   final SecondSubtype instance => {
+  ///       'type': 'SecondSubtype',
+  ///       ..._$SecondSubtypeToJson(instance),
+  ///     },
+  /// }
+  /// ```
+  String _createSealedFunctionExpressionBody() {
+    assert(element.isSealed);
 
-          final keyValuePair = '$keyExpression: $maybeQuestion$valueExpression';
-          return '        $keyValuePair,\n';
-        }),
-      )
-      ..writeln('};');
+    final implementations = sealedSubClasses(element);
 
-    yield buffer.toString();
+    final discriminator = config.unionDiscriminator;
+
+    String buildSingleImpl(ClassElement impl) {
+      final originalName = impl.name!;
+
+      final unionName = encodedFieldName(config.unionRename, originalName);
+
+      return '''
+  final $originalName instance => {
+    '$discriminator': '$unionName',
+    ...${classPrefix(impl)}ToJson(instance),
+  },
+''';
+    }
+
+    final buffer = StringBuffer()
+      ..writeln('switch (instance) {')
+      ..writeAll(implementations.map(buildSingleImpl))
+      ..writeln('}');
+
+    return buffer.toString();
+  }
+
+  /// Creates expression body for a function that serializes a class.
+  ///
+  /// For example:
+  /// ```dart
+  /// <String, dynamic>{
+  ///   'exampleField': instance.exampleField,
+  /// }
+  /// ```
+  String _createFieldMapFunctionExpressionBody(
+    Set<FieldElement> accessibleFields,
+  ) {
+    String buildSingleField(FieldElement field) {
+      final access = _fieldAccess(field);
+
+      final keyExpression = safeNameAccess(field);
+      final valueExpression = _serializeField(field, access);
+
+      final maybeQuestion = _canWriteJsonWithoutNullCheck(field) ? '' : '?';
+
+      return '$keyExpression: $maybeQuestion$valueExpression,';
+    }
+
+    final buffer = StringBuffer()
+      ..writeln('<String, dynamic>{')
+      ..writeAll(accessibleFields.map(buildSingleField))
+      ..writeln('}');
+
+    return buffer.toString();
   }
 
   void _writeGenericArgumentFactories(StringBuffer buffer) {