+++ json_serializable/lib/src/utils.dart
@@ -9,11 +9,17 @@ import 'package:json_annotation/json_annotation.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:source_helper/source_helper.dart';
 
+import 'settings.dart';
 import 'shared_checkers.dart';
 import 'type_helpers/config_types.dart';
 
 const _jsonKeyChecker = TypeChecker.typeNamed(
   JsonKey,
+  inPackage: 'json_annotation',
+);
+
+const _jsonSerializableChecker = TypeChecker.typeNamed(
+  JsonSerializable,
   inPackage: 'json_annotation',
 );
 
@@ -74,6 +80,8 @@ JsonSerializable _valueForAnnotation(ConstantReader reader) => JsonSerializable(
       reader.read('genericArgumentFactories').literalValue as bool?,
   ignoreUnannotated: reader.read('ignoreUnannotated').literalValue as bool?,
   includeIfNull: reader.read('includeIfNull').literalValue as bool?,
+  unionDiscriminator: reader.read('unionDiscriminator').literalValue as String?,
+  unionRename: readEnum(reader.read('unionRename'), FieldRename.values),
 );
 
 /// Returns a [ClassConfig] with values from the [JsonSerializable]
@@ -127,6 +135,9 @@ ClassConfig mergeConfig(
     includeIfNull: annotation.includeIfNull ?? config.includeIfNull,
     ctorParams: ctorParams,
     converters: converters.isNull ? const [] : converters.listValue,
+    unionDiscriminator:
+        annotation.unionDiscriminator ?? config.unionDiscriminator,
+    unionRename: annotation.unionRename ?? config.unionRename,
   );
 }
 
@@ -167,6 +178,61 @@ ConstructorElement constructorByName(ClassElement classElement, String name) {
   return ctor;
 }
 
+/// Given a [ClassElement] that is a sealed class, returns all the
+/// direct subclasses of the given sealed class, excluding any
+/// indirect subclasses (ie. subclasses of subclasses).
+///
+/// Otherwise, returns an empty iterable.
+Iterable<ClassElement> sealedSubClasses(ClassElement maybeSealedSuperClass) {
+  if (maybeSealedSuperClass case final sc when sc.isSealed) {
+    return LibraryReader(
+      sc.library,
+    ).allElements.whereType<ClassElement>().where(
+      (e) => e.interfaces.contains(sc.thisType) || e.supertype?.element == sc,
+    );
+  }
+
+  return const Iterable<ClassElement>.empty();
+}
+
+/// Given a [ClassElement] that is a subclass of sealed classes, returns
+/// all of the sealed superclasses, including all indirect superclasses
+/// (ie. superclasses of superclasses)
+///
+/// Otherwise, returns an empty iterable.
+Iterable<ClassElement> sealedSuperClasses(
+  ClassElement maybeSealedImplementation,
+) => maybeSealedImplementation.allSupertypes
+    .map((type) => type.element)
+    .whereType<ClassElement>()
+    .where((element) => element.isSealed);
+
+/// Given a [ClassElement] that is annotated with `@JsonSerializable`, returns
+/// the annotation config merged with build runner config and defaults.
+///
+/// Otherwise, returns `null`.
+ClassConfig? jsonSerializableConfig(
+  ClassElement maybeAnnotatedElement,
+  Settings generator,
+) {
+  final maybeSuperAnnotation = _jsonSerializableChecker.firstAnnotationOfExact(
+    maybeAnnotatedElement,
+    throwOnUnresolved: false,
+  );
+
+  if (maybeSuperAnnotation case final superAnnotation?) {
+    final annotationReader = ConstantReader(superAnnotation);
+
+    return mergeConfig(
+      generator.config,
+      annotationReader,
+      classElement: maybeAnnotatedElement,
+    );
+  }
+
+  return null;
+}
+
 /// If [targetType] is an enum, returns the [FieldElement] instances associated
 /// with its values.
 ///
@@ -200,6 +266,8 @@ String encodedFieldName(FieldRename fieldRename, String declaredName) =>
       FieldRename.kebab => declaredName.kebab,
       FieldRename.pascal => declaredName.pascal,
     };
+
+String classPrefix(ClassElement element) => '_\$${element.name!.nonPrivate}';
 
 /// Return the Dart code presentation for the given [type].
 ///