+++ json_serializable/test/test_sources/test_sources.dart
@@ -24,6 +24,8 @@ class ConfigurationImplicitDefaults {
   ignoreUnannotated: false,
   includeIfNull: true,
   genericArgumentFactories: false,
+  unionDiscriminator: 'type',
+  unionRename: FieldRename.none,
 )
 class ConfigurationExplicitDefaults {
   int? field;